requirements.txt
pyserial
pyhidapi